Python script for cdp nei details
This Python script is used to display the output of show cdp nei on Cisco devices. In the 'myswitches' text document is the IP's of the switches, routers you want to fetch the info from. It will create a text file with cdp info from all IP's in the 'myswitches' file.